+++ src/metric/LCS.ts
@@ -63,7 +63,7 @@ export class LCSMetric extends Metric< LCSRaw > {
     ) : MetricCompute< LCSRaw > {
         // Get two reusable arrays from the Pool for the DP rows
         const len = m + 1;
-        const [ prev, curr ] = Pool.acquireMany< Uint16Array >( 'uint16', [ len, len ] );
+        const [ prev, curr ] = Pool.acquireMany< Int32Array >( 'int32', [ len, len ] );
 
         try {
             // Initialize the first row to zeros
@@ -98,8 +98,8 @@ export class LCSMetric extends Metric< LCSRaw > {
             };
         } finally {
             // Release arrays back to the pool
-            Pool.release( 'uint16', prev, len );
-            Pool.release( 'uint16', curr, len );
+            Pool.release( 'int32', prev, len );
+            Pool.release( 'int32', curr, len );
         }
     }
