+++ src/metric/NeedlemanWunsch.ts
@@ -63,7 +63,7 @@ export class NeedlemanWunschDistance extends Metric< NeedlemanRaw > {
 
         // Get two reusable arrays from the Pool for the DP rows
         const len = m + 1;
-        const [ prev, curr ] = Pool.acquireMany< Uint16Array >( 'uint16', [ len, len ] );
+        const [ prev, curr ] = Pool.acquireMany< Int32Array >( 'int32', [ len, len ] );
 
         try {
             // Initialize the first row (gap penalties)
@@ -105,8 +105,8 @@ export class NeedlemanWunschDistance extends Metric< NeedlemanRaw > {
             };
         } finally {
             // Release arrays back to the pool
-            Pool.release( 'uint16', prev, len );
-            Pool.release( 'uint16', curr, len );
+            Pool.release( 'int32', prev, len );
+            Pool.release( 'int32', curr, len );
         }
     }
